update operators and samplers dirs
