+++ src/operators/operators.jl
@@ -1,6 +1,21 @@
 include("utils.jl")
+
 include("initializer.jl")
-include("selection.jl")
-include("crossover.jl")
-include("mutation.jl")
+include("sampling/random.jl")
+include("sampling/grid.jl")
+include("sampling/latinhypercube.jl")
+include("sampling/searchspaces.jl")
+
+include("selection/selection.jl")
+
+# crossover
+include("crossover/order.jl")
+include("crossover/sbx.jl")
+include("crossover/uniform.jl")
+
+include("mutation/bitflip.jl")
+include("mutation/polynomial.jl")
+include("mutation/slight.jl")
+include("mutation/mutation.jl")
+
 include("environmental_selection.jl")
